# Remove usage of `unsafe.Pointer` in `Map` wherever possible