--- a/export_test.go
+++ b/export_test.go
@@ -9,7 +9,7 @@ const (
 )
 
 type (
-	BucketPadded = bucketPadded
+	BucketPadded[K comparable, V any] = bucketPadded[K, V]
 )
 
 func EnableAssertions() {