@puzpuzpuz 2025-04-01
TBH I find the old code readable enough, but I'm biased. Anyhow, I more than fine to switch to the atomic types.