@PapaCharlie 2025-04-01
That's fair! And to be clear, I was in no way implying that the code wasn't legible at the time. We just have access to better syntax now. Thanks for merging this!